More common send/receive: Header partly common Incorrect size for sending packets to 625XT

+++ Device/Gh505Packet.cs
@@ -225,7 +225,7 @@ namespace ZoneFiveSoftware.SportTracks.Device.Globalsat
             int startOffset = offset;
             offset += this.Write(offset, trackFile.StartTime);
 
-            offset += this.Write(offset, trackFile.TrackPointCount);
+            offset += this.Write(offset, (Int16)trackFile.TrackPointCount);
             int totalTimeSecondsTimes10 = ToGlobTime(trackFile.TotalTime.TotalSeconds);
             offset += this.Write32(offset, totalTimeSecondsTimes10);
             offset += this.Write32(offset, (Int32)trackFile.TotalDistanceMeters);
